requirements.txt
pandas==2.2.3
pytest==8.1.1
pytest-bdd==7.2.0
allure-pytest-bdd==2.13.5
pytest-jira==0.3.21
pyhamcrest==2.1.0
boto3==1.20.23
python-dotenv==1.0.1
pytest-xdist==3.6.1
pytest-retry==1.6.3
requests==2.32.3
requests-oauthlib==2.0.0
teradatasql==20.0.0.15
jira==3.8.0
cryptography==43.0.1
XlsxWriter==3.2.0
customtkinter==5.2.2
sqlparse==0.5.1
teradata==15.10.0.21
